DetailPage : Design our course detail page.

+++ lib/common/widgets/base_text_widget.dart
@@ -7,7 +7,7 @@ import '../values/colors.dart';
 Widget reuseableText(
   String text, {
   Color color = AppColors.primaryText,
-  int fontSize = 16,
+  double fontSize = 16,
   FontWeight fontWeight = FontWeight.bold,
 }) {
   return Text(
